1st copy
Alright it’s 2 in the morning we created the project, generated our
app, and began working on structure of the main layout. This included
the About/Contact section on the left side. The navigation bar with
simple scss animations. Tried out some fonts but none of those are for
certain.

TODOS:
- More animations on the nav bar, make it really fun to use and
everything rather obvious.
-build grid and figure out the angular way of looping photos.
- how create links for individual photos and their description

*****F*U*T*U*R*E**********
Flickr Api