created the base for navigating through the galleries
Easy way to differentiate sections and loop photos using AngularJS.
Today we created the plan in which creating the app will be really easy
and fun.

TODOS:
- Create looping systems once the user clicks the image where they’ll
be lead to a large description followed by a close up of the image
-Set it up so that after the image the user is lead to the next image
in the sequence.
- the user should be able to click back on the ‘X’ button to the area
where the image is located
- DESIGN DESIGN DESIGN!!!!!!
- being able to limit the amount of images originally showed to 4 lines
and have the user click on the arrow button to reveal more.